Conflicts:
	Core/MIPS/ARM/ArmJit.h
	Core/MIPS/ARM/ArmRegCacheFPU.cpp
	Core/MIPS/ARM/ArmRegCacheFPU.h
	unittest/UnitTest.cpp

+++ Core/FileSystems/BlockDevices.h
@@ -23,7 +23,7 @@
 // The ISOFileSystemReader reads from a BlockDevice, so it automatically works
 // with CISO images.
 
-#include "../../Globals.h"
+#include "Common/CommonTypes.h"
 #include "Core/ELF/PBPReader.h"
 
 class BlockDevice
@@ -63,7 +63,7 @@ public:
 
 private:
 	FILE *f;
-	size_t filesize;
+	u64 filesize;
 };
